+++ CHANGELOG.md
@@ -1,4 +1,25 @@
 # CHANGELOG
+## [1.1.0]
+### Added
+- Support for Imou home security device integration
+- Support for multiple lens camera integration
+- Support for integrated option configuration
+- Support for binary sensor entity type
+- Support for Chinese translation of entity status
+
+### Changed
+- Fixed some bugs
+- Optimized interaction logic with the platform, offline devices will no longer request updates
+- Other code optimizations
+
+## [1.0.1]
+### Added
+- Test case
+- Github action
+- Pre-commit hook
+
+### Changed
+- Code optimization
 
 ## [1.0.0]